Small Changes
Added Animator to all the Buttons

Added the functions to the GameOver buttons (although they dont reset
all the screen and values properly)

Added parents to buttons where necessary so that they animate anchored
to their center, instead of to sides or corners.

Locked to Portrait.

Removed 1 button from the Character_Select screen, and added a
horizontal layout group so that button sizes change with screen sizes
(was not adapting before).